Enhanced error handling and improved code readability
In this commit, error handling was enhanced in the admin component with the introduction of an error property to display error messages related to product creation. Similarly, the ViewChild reference was improved by using a more unique and specific property name to ease identification. This change introduces an AfterViewInit lifecycle hook to fix initialization issues with ViewChild. Removed some commented out code for more readability and moved initialization of some event listeners to the AfterViewInit lifecycle method and removed them in the OnDestroy method. Fixed some misnomers in checkout component, and updated README and instructions for local setup and admin access.